+++ android/src/main/kotlin/com/sap/gigya_flutter_plugin/GigyaFlutterPlugin.kt
@@ -1,9 +1,16 @@
 package com.sap.gigya_flutter_plugin
 
 import android.app.Application
+import androidx.activity.ComponentActivity
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.IntentSenderRequest
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.NonNull
+import com.gigya.android.sdk.Gigya
 import com.gigya.android.sdk.account.models.GigyaAccount
 import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
@@ -11,7 +18,7 @@ import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 
 /** GigyaFlutterPlugin */
-class GigyaFlutterPlugin : FlutterPlugin, MethodCallHandler {
+class GigyaFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
     companion object {
         private lateinit var sdk: GigyaSDKWrapper<*>
@@ -22,9 +29,10 @@ class GigyaFlutterPlugin : FlutterPlugin, MethodCallHandler {
         }
     }
 
-
     /// Main communication method channel.
     private lateinit var channel: MethodChannel
+
+    private var fidoResultHandler: ActivityResultLauncher<IntentSenderRequest>? = null
 
     private var _messenger: BinaryMessenger? = null
 
@@ -39,6 +47,7 @@ class GigyaFlutterPlugin : FlutterPlugin, MethodCallHandler {
         channel.setMethodCallHandler(null)
         _messenger = null
     }
+
 
     /// Main channel call handler. Will initiate native wrapper.
     override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
@@ -61,10 +70,49 @@ class GigyaFlutterPlugin : FlutterPlugin, MethodCallHandler {
             "resolveSetAccount" -> sdk.resolveSetAccount(call.arguments, result)
             "forgotPassword" -> sdk.forgotPassword(call.arguments, result)
             "initSdk" -> sdk.initSdk(call.arguments, result)
+            "getSession" -> sdk.getSession(result)
             "setSession" -> sdk.setSession(call.arguments, result)
             "sso" -> sdk.sso(call.arguments, result)
+            "webAuthnLogin" -> {
+                fidoResultHandler?.let {
+                    sdk.webAuthnLogin(it, result)
+                }
+            }
+            "webAuthnRegister" -> {
+                fidoResultHandler?.let {
+                    sdk.webAuthnRegister(it, result)
+                }
+            }
+            "webAuthnRevoke" -> sdk.webAuthnRevoke(result)
             else -> result.notImplemented()
         }
+    }
+
+    override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+        val activity = binding.activity
+        if (activity is ComponentActivity) {
+            fidoResultHandler = activity.registerForActivityResult(
+                ActivityResultContracts.StartIntentSenderForResult()
+            ) { activityResult ->
+                val extras =
+                    activityResult.data?.extras?.keySet()
+                        ?.map { "$it: ${activity.intent.extras?.get(it)}" }
+                        ?.joinToString { it }
+                Gigya.getInstance().WebAuthn().handleFidoResult(activityResult)
+            }
+        }
+    }
+
+    override fun onDetachedFromActivityForConfigChanges() {
+
+    }
+
+    override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
+
+    }
+
+    override fun onDetachedFromActivity() {
+        fidoResultHandler = null
     }