+++ example/android/app/src/main/kotlin/com/sap/gigya_flutter_plugin_example/MainActivity.kt
@@ -1,5 +1,36 @@
 package com.sap.gigya_flutter_plugin_example
 
-import io.flutter.embedding.android.FlutterActivity
+import android.content.pm.PackageManager
+import android.os.Bundle
+import android.util.Base64
+import android.util.Log
+import io.flutter.embedding.android.FlutterFragmentActivity
+import java.security.MessageDigest
+import java.security.NoSuchAlgorithmException
 
-class MainActivity : FlutterActivity()
+class MainActivity : FlutterFragmentActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        getSignature()
+    }
+
+    // Call function to get the application signature used for Facebook login or Fido authentication.
+    private fun getSignature() {
+        try {
+            val info = packageManager.getPackageInfo(packageName, PackageManager.GET_SIGNATURES)
+            for (signature in info.signatures) {
+                val md: MessageDigest = MessageDigest.getInstance("SHA256")
+                md.update(signature.toByteArray())
+                Log.e(
+                    "MY KEY HASH:", Base64.encodeToString(
+                        md.digest(),
+                        Base64.URL_SAFE or Base64.NO_PADDING or Base64.NO_WRAP
+                    )
+                )
+            }
+        } catch (e: PackageManager.NameNotFoundException) {
+        } catch (e: NoSuchAlgorithmException) {
+        }
+    }
+}